Rename popver_calculate_position to calculate_popover_position.
